Three new sections add to BE of HP

+++ wp-content/plugins/formidable-pro/classes/models/FrmProContent.php
@@ -1,5 +1,9 @@
 <?php
 
+if ( ! defined( 'ABSPATH' ) ) {
+	die( 'You are not allowed to call this page directly.' );
+}
+
 /**
  * Class FrmProContent
  */
@@ -83,6 +87,7 @@ class FrmProContent {
 
 		if ( ! $foreach && ! $conditional && isset( $atts['show'] ) && ( $atts['show'] == 'field_label' || $atts['show'] == 'description' ) ) {
 			// get the field label or description and return before any other checking
+			$field = apply_filters( 'frm_field_object_for_shortcode', $field );
 			$replace_with = ( $atts['show'] == 'field_label' ) ? $field->name : $field->description;
 			$content = str_replace( $shortcodes[0][ $short_key ], $replace_with, $content );
 			return;
@@ -120,6 +125,11 @@ class FrmProContent {
 			if ( ! $keep_array && $field->type != 'file' ) {
 				$replace_with = FrmAppHelper::array_flatten( $replace_with );
 				$replace_with = array_filter( $replace_with, array( 'FrmProContent', 'is_not_empty' ) );
+
+				if ( ! isset( $atts['sep'] ) && FrmProImages::has_images_options_in_html( $replace_with ) ) {
+					$sep = '';
+				}
+
 				$replace_with = implode( $sep, $replace_with );
 			} else if ( empty( $replace_with ) ) {
 				$replace_with = '';
@@ -135,9 +145,10 @@ class FrmProContent {
 			self::check_conditional_shortcode( $content, $replace_with, $atts, $tag, 'if', array( 'field' => $field ) );
 		} else {
 			if ( empty( $replace_with ) && $replace_with != '0' ) {
-				$replace_with = '';
-				if ( $field->type == 'number' ) {
-					$replace_with = '0';
+				if ( isset( $atts['default'] ) ) {
+					$replace_with = $atts['default'];
+				} else {
+					$replace_with = '';
 				}
 			} else {
 				$replace_with = FrmFieldsHelper::get_display_value( $replace_with, $field, $atts );
@@ -180,6 +191,18 @@ class FrmProContent {
 	 */
 	public static function is_not_empty( $val ) {
 		return $val !== '';
+	}
+
+	/**
+	 * Filter out entry_number shortcode when we have the entry position in the view
+	 *
+	 * @since 4.03.01
+	 */
+	public static function replace_entry_position_shortcode( $entry_args, $args, &$content ) {
+		preg_match_all( "/\[(if )?(entry_position)\b(.*?)(?:(\/))?\](?:(.+?)\[\/\2\])?/s", $content, $shortcodes, PREG_PATTERN_ORDER );
+		foreach ( $shortcodes[0] as $short_key => $tag ) {
+			self::replace_single_shortcode( $shortcodes, $short_key, $tag, $entry_args['entry'], $entry_args['view'], $args, $content );
+		}
 	}
 
 	public static function replace_calendar_date_shortcode( $content, $date ) {
@@ -242,8 +265,8 @@ class FrmProContent {
 	 */
 	public static function get_pretty_url( $atts ) {
 		global $post;
-		$base_url = untrailingslashit( $post ? get_permalink( $post->ID ) : $_SERVER['REQUEST_URI'] );
-		if ( ! is_front_page() && self::rewriting_on() ) {
+		$base_url = untrailingslashit( $post ? get_permalink( $post->ID ) : FrmAppHelper::get_server_value( 'REQUEST_URI' ) );
+		if ( ! is_front_page() && self::rewriting_on() && strpos( $base_url, '?' ) === false ) {
 			$url = $base_url . '/' . $atts['param'] . '/' . $atts['param_value'];
 		} else {
 			$url = esc_url_raw( add_query_arg( $atts['param'], $atts['param_value'], $base_url ) );
@@ -271,6 +294,7 @@ class FrmProContent {
 	/**
 	 * This is a workaround for a bug in WordPress Core
 	 * https://core.trac.wordpress.org/ticket/23867
+	 *
 	 * @since 2.2.10
 	 */
 	public static function fix_home_page_query( $query ) {
@@ -293,7 +317,25 @@ class FrmProContent {
 	private static function get_rewrite_params() {
 		global $wpdb;
 		$params = FrmDb::get_col( $wpdb->postmeta, array( 'meta_key' => 'frm_param' ), 'meta_value' );
+		$params = self::remove_reserved_words( $params );
 		return array_filter( array_unique( $params ) );
+	}
+
+	/**
+	 * Removes reserved words from a list of params.
+	 *
+	 * @param array $params A list of params.
+	 *
+	 * @return array The list of params with reserved words removed.
+	 */
+	private static function remove_reserved_words( $params ) {
+		if ( ! is_callable( 'FrmFormsHelper::reserved_words' ) ) {
+			return $params;
+		}
+
+		$reserved_words = FrmFormsHelper::reserved_words();
+
+		return array_diff( $params, $reserved_words );
 	}
 
 	public static function do_shortcode_editlink( &$content, $atts, $shortcodes, $short_key, $args ) {
@@ -318,7 +360,7 @@ class FrmProContent {
 			if ( $args['entry']->post_id ) {
 				$replace_with = get_edit_post_link( $args['entry']->post_id );
 			} else if ( current_user_can('frm_edit_entries') ) {
-				$replace_with = admin_url( 'admin.php?page=formidable-entries&frm_action=edit&id=' . $args['entry']->id );
+				$replace_with = FrmProEntry::admin_edit_link( $args['entry']->id );
 			}
 
 			if ( ! empty( $replace_with ) ) {
@@ -375,7 +417,9 @@ class FrmProContent {
 			self::check_conditional_shortcode( $content, $args['entry']->{$args['tag']}, $atts, $args['tag'] );
 		} else {
 			if ( isset( $atts['time_ago'] ) ) {
-				$date = FrmAppHelper::human_time_diff( strtotime( $args['entry']->{$args['tag']} ), '', absint( $atts['time_ago'] ) );
+				// $time_ago values can include 1-7, as well as d,w,h,s and their extended day,week,hour,second names as well.
+				$time_ago = is_numeric( $atts['time_ago'] ) ? absint( $atts['time_ago'] ) : sanitize_key( $atts['time_ago'] );
+				$date     = FrmAppHelper::human_time_diff( strtotime( $args['entry']->{$args['tag']} ), '', $time_ago );
 			} else {
 				$date = FrmAppHelper::get_formatted_time( $args['entry']->{$args['tag']}, $atts['format'], $time_format );
 			}
@@ -578,7 +622,18 @@ class FrmProContent {
 					$replace_with = '';
 				}
 			}
-		} elseif ( ( $field && $field->type == 'user_id' ) || in_array( $tag, array( 'updated_by', 'created_by' ) ) ) {
+
+			// Get the linked field to properly evaluate conditions
+			if ( $replace_with !== '' && isset( $atts['show'] ) && ! empty( $atts['show'] ) ) {
+				$show_field = FrmField::getOne( $atts['show'] );
+				if ( $show_field && in_array( $show_field->type, array( 'time', 'date', 'user_id' ) ) ) {
+					$field = $show_field;
+					unset( $atts['show'] );
+				}
+			}
+		}
+
+		if ( ( $field && $field->type == 'user_id' ) || in_array( $tag, array( 'updated_by', 'created_by' ) ) ) {
 			// check if conditional is for current user
 			if ( isset( $atts['equals'] ) && $atts['equals'] == 'current' ) {
 				$atts['equals'] = get_current_user_id();
@@ -601,7 +656,7 @@ class FrmProContent {
 					if ( strtolower( $atts[ $att_name ] ) == 'now' ) {
 						$atts[ $att_name ] = FrmProAppHelper::get_date( 'H:i' );
 					} else {
-						$atts[ $att_name ] = date( 'H:i', strtotime( $atts[ $att_name ] ) );
+						$atts[ $att_name ] = gmdate( 'H:i', strtotime( $atts[ $att_name ] ) );
 					}
 
 					if ( ! $formatted_time ) {
@@ -609,6 +664,11 @@ class FrmProContent {
 						$formatted_time = true;
 					}
 				}
+			}
+		} else {
+			// Compare properly with &.
+			if ( is_callable( 'FrmAppHelper::decode_specialchars' ) ) {
+				FrmAppHelper::decode_specialchars( $replace_with );
 			}
 		}
 
@@ -629,7 +689,7 @@ class FrmProContent {
 				} elseif ( $atts[ $att_name ] == 'NOW' ) {
 					$atts[ $att_name ] = FrmProAppHelper::get_date( 'Y-m-d' );
 				} else {
-					$atts[ $att_name ] = date( 'Y-m-d', strtotime( $atts[ $att_name ] ) );
+					$atts[ $att_name ] = gmdate( 'Y-m-d', strtotime( $atts[ $att_name ] ) );
 				}
 			}
 			unset( $att_name );
@@ -645,7 +705,7 @@ class FrmProContent {
 
 		$compare = strtolower( $compare );
 		if ( strpos( $compare, 'like' ) === false ) {
-			$where_val = date( 'Y-m-d H:i:s', strtotime( $where_val ) );
+			$where_val = gmdate( 'Y-m-d H:i:s', strtotime( $where_val ) );
 
 			// If using less than or equal to, set the time to the end of the day
 			if ( $compare == '<=' || $compare == 'less_than' ) {
@@ -740,9 +800,7 @@ class FrmProContent {
 	}
 
 	private static function eval_less_than_condition( $atts, &$field_value ) {
-		if ( $field_value < $atts['less_than'] ) {
-			// Condition is true
-		} else {
+		if ( $field_value >= $atts['less_than'] ) {
 			// Condition is false
 			$field_value = '';
 		}
@@ -756,9 +814,7 @@ class FrmProContent {
 	}
 
 	private static function eval_greater_than_condition( $atts, &$field_value ) {
-		if ( $field_value > $atts['greater_than'] ) {
-			// Condition is true
-		} else {
+		if ( $field_value <= $atts['greater_than'] ) {
 			// Condition is false
 			$field_value = '';
 		}
@@ -810,9 +866,15 @@ class FrmProContent {
 			return FrmAppHelper::truncate( $replace_with, (int) $atts['truncate'], 3, $more_link_text );
 		}
 
-		$clean_text = wp_strip_all_tags( $replace_with );
+		if ( empty( $replace_with ) ) {
+			return $replace_with;
+		}
+
+		$clean_text = trim( wp_strip_all_tags( $replace_with ) );
 		$part_one = FrmAppHelper::truncate( $clean_text, (int) $atts['truncate'], 3, '' );
-		$part_two = str_replace( $part_one, '', $clean_text );
+		$pos      = strpos( $clean_text, $part_one );
+		// Only replace the first occurance of the string.
+		$part_two = substr_replace( $clean_text, '', $pos, strlen( $part_one ) );
 
 		if ( ! empty( $part_two ) ) {
 			$replace_with = $part_one . '<a href="#" onclick="jQuery(this).next().css(\'display\', \'inline\');jQuery(this).css(\'display\', \'none\');return false;" class="frm_text_exposed_show"> ' . $more_link_text . '</a><span style="display:none;">' . $part_two . '</span>';
